% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/age_cat.r
\name{age_cat}
\alias{age_cat}
\title{Categorize a continuous age variable into common categories.}
\usage{
age_cat(data, agevar, groups = "census", factor = TRUE)
}
\arguments{
\item{data}{Data Frame with the age variable to categorize.}

\item{agevar}{Name of the age variable.}

\item{groups}{Categorization Scheme: census, census65,  five, or ten.}

\item{factor}{Return a factor or not.  Defaults to TRUE}
}
\value{
A factor, unless otherwise specified, with recoded \code{agevar}.
}
\description{
Categorize a continuous age variable into common categories.
}
\examples{
age.cat()
}
\keyword{age,}
\keyword{recode}

