% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ann_gr.r
\name{ann_gr}
\alias{ann_gr}
\title{Calculates annualized growth rates.}
\usage{
ann_gr(begin, end, n)
}
\arguments{
\item{begin}{Value at the beginning of the period.}

\item{end}{Value at the end of the period.}

\item{n}{Number of years from beginning to end}
}
\description{
Calculates annualized growth rates.
}
\examples{
ann.gr(4300,7950, 10)
}
\keyword{annualized}
\keyword{growth}
\keyword{rate,}

