\name{codemog-package}
\alias{codemog-package}
\alias{codemog}
\docType{package}
\title{
Colorado State Demography Office Population Data

}
\description{
This package contains the most recent vintage of population estimates from the Colorado State Demography Office. Full documentation here: <http://rpubs.com/robkemp/42244>
}
\details{
\tabular{ll}{
Package: \tab codemog\cr
Type: \tab Package\cr
Version: \tab 1.0.13\cr
Date: \tab 2014-11-14\cr
License: \tab MIT + LICENSE \cr
}
The version number is in three sections, the first shows the version for the data, the second shows the version for the R functions, and the third contains the last two digits of the vintage year for the estimates (e.g. Vintage 2013 estimates would use .13)
}
\author{
Rob Kemp <robert.kemp@state.co.us>

Maintainer: Rob Kemp <robert.kemp@state.co.us>

}


\keyword{ population data }


