% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/codemog_api.r
\name{codemog_api}
\alias{codemog_api}
\title{Links R to the URL based REST API for the CO Demography Office Databases}
\usage{
codemog_api(datacall = "table", data, db = "c2010", geonum = "108",
  sumlev = NULL, type = "csv", meta = "yes", geography = "geonum")
}
\arguments{
\item{datacall}{table or field based on call, field if specific column, table for all Defaults to table}

\item{data}{number of either a census product table or table and field number based on the datacall}

\item{db}{database c1980, c1990, c2000, c2010 for Census Data, acs0812 or acs0913 for ACS data}

\item{geonum}{a geographic identifier created by 1 followed by a State FIPS and a PLace/County FIPS Defaults to 108}

\item{sumlev}{can be used to call all geographies in a summary level (See Census for Definitions) Defaults to NULL}

\item{type}{can siwtch between JSON and CSV for output format (pretty much use CSV only) Defaults to CSV}

\item{meta}{a command that indicates whether include feild and table meta data in line 2 Defaults to yes}

\item{geography}{is a switch that chooses use of a geonum for one area or a summary level for a specific type Defaults to geonum}
}
\description{
Takes some basic geographic information along with the table/field and
dataset to grab tables from the CO Demography Office cleaned Census
databases.
}

