% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/graphs.R
\docType{data}
\name{codemog_pal}
\alias{codemog_pal}
\title{Colorado State Demography Office Color Palette for ggplot2}
\format{\preformatted{ Named chr [1:6] "#1F497D" "#C0504D" "#4E5758" "#D2D2D2" "#D0D2D3" "#77AB43"
 - attr(*, "names")= chr [1:6] "dkblu" "dkred" "dkgray" "medgray" ...
}}
\usage{
codemog_pal
}
\description{
Custom color palette using a mix of SDO colors and DOLA
Brand Colors from Brand Colorado.
}
\keyword{datasets}

