% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_base}
\alias{county_base}
\title{County Base Industry Analysis Data}
\format{A data frame with nine variables: \code{geonum}, \code{countyfips},
  \code{county}, \code{year}, \code{baseIndustry}, \code{jobs},\code{geonum}, and \code{industry_group}.}
\usage{
county_base

county_base
}
\description{
Jobs by Base Industry for Colorado Counties and the Metro Area in 2014

Jobs by Base Industry for Colorado Counties and the Metro Area in 2014
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset: http://www.github.com/robkemp/codemog

See 'Data Dictionary' vignette for specifics on the dataset: http://www.github.com/robkemp/codemog
}
\keyword{datasets}

