% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_forecast}
\alias{county_forecast}
\title{County by Age and Sex Estimates and Forecasts}
\format{A data frame with nine variables: \code{countyfips},\code{year}, \code{age},
  \code{county},  \code{male Population}, \code{femalePopulation},
  \code{totalPopulation}, and \code{datatype}.}
\usage{
county_forecast
}
\description{
Population Estimates from 1990 to 2013 and population forecasts from 2014 to 2040 for
counties in Colorado.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset
}
\keyword{datasets}

