% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_hist}
\alias{county_hist}
\title{County Historical Population Estimates}
\format{A data frame with nine variables: \code{geonum}, \code{countyfips},
  \code{county}, \code{year}, \code{totalPopulation}, and \code{datatype}.}
\usage{
county_hist
}
\description{
Population Estimates from 1980 to 2009 for the Counties in Colorado.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset
}
\keyword{datasets}

