% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{county_jobs}
\alias{county_jobs}
\title{County Jobs Data}
\format{A data frame with nine variables: \code{geonum}, \code{countyfips},
  \code{sector_id}, \code{year}, \code{sector_name}, \code{jobs},\code{geonum}, and \code{total_flag}.}
\usage{
county_jobs
}
\description{
Jobs by sector from 2001 to 2013, estimates created by the State Demography Office.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset: http://www.github.com/robkemp/codemog
}
\keyword{datasets}

