% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emp_cat.r
\name{emp_cat}
\alias{emp_cat}
\title{Recategorize the ACS Employment status variable (ESR)}
\usage{
emp_cat(data, empvar = "ESR", groups = "basic", factor = TRUE)
}
\arguments{
\item{data}{Data Frame with the education variable to categorize.}

\item{groups}{Categorization Scheme: basic, military, lf Defaults to basic}

\item{factor}{Return a factor or not.  Defaults to TRUE}

\item{employment}{var Name of the education variable.  Defaults to ESR}
}
\value{
A factor, unless otherwise specified, with recoded \code{edvar}.
}
\description{
Recategorize the ACS Employment status variable (ESR)
}
\examples{
emp_cat()
}
\keyword{employment,}
\keyword{recode}

