% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/graphs.R
\name{ms_census_age}
\alias{ms_census_age}
\title{Creates a \code{ggplot2} chart of the Census Age Distribution in 2000 and 2010.}
\usage{
ms_census_age(fips, state = "08")
}
\arguments{
\item{fips}{A Place (or County) FIPS number}

\item{state}{The initial State using a State FIPS (08=CO) Defaults to 08.}
}
\description{
Takes a place and then  creates a plot of the 2000 and 2010  data in \code{ggplot2}.
Note: Requires dplyr, ggplot2, ggthemes, scales, grid,
reshape2,tidur, stringi R packages.
}

