% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/graphs.R
\name{ms_ed}
\alias{ms_ed}
\title{Creates a \code{ggplot2} chart of the educational attainment comparison}
\usage{
ms_ed(fips, state = "08", fips2 = "", state2 = "08")
}
\arguments{
\item{fips}{A Place (or County) FIPS number}

\item{state}{The initial State using a State FIPS (08=CO) Defaults to 08.}

\item{fips2}{A Place (or County) FIPS number for comparison Defaults to ""}

\item{state2}{The comparison place's State using a State FIPS (08=CO) Defaults to 08.}
}
\description{
Takes two places (using fips numbers) and then  creates a plot of the data in \code{ggplot2}.
Note: Requires dplyr, ggplot2, ggthemes, scales, grid,
reshape2,tidur, stringi R packages.
}

