% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{muni_est}
\alias{muni_est}
\title{Municipal Estimates}
\format{A data frame with nine variables: \code{geonum}, \code{placefips},
  \code{municipality}, \code{year}, \code{groupquartersPopulation}, \code{householdPopulation},\code{occupiedHousehold},
   \code{totalHousingUnits}, and \code{totalPopulation}.}
\usage{
muni_est
}
\description{
Population Estimates from 2010 to 2013 for the Municipalities in Colorado.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset
}
\keyword{datasets}

