% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{muni_hist}
\alias{muni_hist}
\title{Municipal Historical Population Estimates}
\format{A data frame with nine variables: \code{geonum}, \code{placefips},
  \code{municipality}, \code{year}, \code{totalPopulation}, and \code{countyfips}.}
\usage{
muni_hist
}
\description{
Population Estimates from 1980 to 2009 for the Municipalities in Colorado.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset
}
\keyword{datasets}

