% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphs.R
\name{muni_ts_chart}
\alias{muni_ts_chart}
\title{Creates a \code{ggplot2} chart of the population for a CO municipality}
\usage{
muni_ts_chart(fips, beginyear = 1990, base = 12)
}
\arguments{
\item{fips}{The County FIPS number (without leading Zeros)}

\item{beginyear}{The first year in the timeseries Defaults to 1990.}

\item{base}{Base font size.}

\item{endyear}{The first year in the timeseries Defaults to 2013.}
}
\description{
Takes some basic input on the timeperiod and county then creates a
plot of the data in \code{ggplot2}.  Similar to the muni_ts_data()
function.  Can create timeseries from 1990 to 2040 (beyond 2013 are
forecasts).
Note: Requires dplyr, ggplot2, ggthemes, scales, and grid R packages.
}

