% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{muni_win_est}
\alias{muni_win_est}
\title{Municipalities within County Estimates}
\format{A data frame with nine variables: \code{geonum}, \code{countyfips},
\code{placefips}, \code{municipality}, \code{year}, \code{groupquartersPopulation},
\code{householdPopulation},\code{occupiedHousehold}, \code{totalHousingUnits},
and \code{totalPopulation}.}
\usage{
muni_win_est
}
\description{
Population Estimates from 2010 to 2013 for the municipalities within counties in
Colorado.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset
}
\keyword{datasets}

