% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{muni_win_hist}
\alias{muni_win_hist}
\title{Municipal parts within counties Historical Population Estimates}
\format{A data frame with nine variables: \code{geonum}, \code{placefips},\code{county_name},
  \code{municipality}, \code{year}, \code{totalPopulation}, and \code{countyfips}.}
\usage{
muni_win_hist
}
\description{
Population Estimates from 1980 to 2009 for the Municipalities within counties in Colorado.
}
\details{
See 'Data Dictionary' vignette for specifics on the dataset
}
\keyword{datasets}

