% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/graphs.R
\name{county_ts_chart}
\alias{county_ts_chart}
\title{Creates a \code{ggplot2} chart of the population for a CO county}
\usage{
county_ts_chart(fips, beginyear = 1990, endyear = 2013, base = 12)
}
\arguments{
\item{fips}{The County FIPS number (without leading Zeros)}

\item{beginyear}{The first year in the timeseries Defaults to 1990.}

\item{endyear}{The first year in the timeseries Defaults to 2013.}

\item{base}{Base font size.}
}
\description{
Takes some basic input on the timeperiod and county then creates a
plot of the data in \code{ggplot2}.  Similar to the county_ts_data()
function.  Can create timeseries from 1990 to 2040 (beyond 2013 are
forecasts).
Note: Requires dplyr, ggplot2, ggthemes, scales, and grid R packages.
}

