% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/graphs.R
\name{theme_codemog}
\alias{theme_codemog}
\title{Colorado State Demography Office ggplot2 Theme}
\usage{
theme_codemog(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Plot text font family.}
}
\description{
Custom \code{ggplot2} theme that borrows heavily from the
\code{theme_fivethirtyeight()} in ggthemes.
}

